Extremely crude easy-smt replacement
I tried to do this as fast as I could, not trying to clean anything up
after the transition, and just get the replacement done as quickly as
possible. (And it still took me about 3 hours.) The main work, of
course, is replacing all the raw string manipulation (mainly `format!`
calls) with S-expression builder calls.