Start conversion to easy-smt (#45)
* Extremely crude easy-smt replacement
I tried to do this as fast as I could, not trying to clean anything up
after the transition, and just get the replacement done as quickly as
possible. (And it still took me about 3 hours.) The main work, of
course, is replacing all the raw string manipulation (mainly `format!`
calls) with S-expression builder calls.
* Try attempting to convert clz encoding...
There is a lot of typing to be done here!
* Try pre-declaring some atoms
* Add a `declare` helper
* Temporarily disable encoded ops
I'm trying to get to a place where the crate compiles, even though the
encoded ops don't work at all.
* Satisfy the borrow checker
* Try converting one rev function
* Add a simple `assume` helper
* Update to latest easy-smt
Now using the builder to construct contexts. And there's a cool new
`numeral` helper for building number literals of all kinds!
* Add back `rev` with one working function & test
* More reusable test framework
* Add encoding files from @mpardesh
* Minor Python cleanup
* Generate easy-smt code for declarations
* Convert assertions to easy-smt
* Slight Python cleanup
* Bind Rust variables for SMT variable names (atoms)
* Slightly simplify some generated Rust
* Finish test for rev1
* Fix variable names
Apparently, `findall` returns a list of strings (not tuples) when there
is only one subgroup.
* Use generated code for rev1
* Generate code for remaining rev*
* Re-enable calls to `rev*`
* Rename cls* SMT files for consistency
To match clz*.
* Clean up test SMT declarations
* Use proper convention for return variables
* Split out `a64cls32` from `cls32` SMT file
Quoth @mpardesh on Slack:
> also note that the cl32 file encodes cls for 32 bits and the a64 version
> after a bunch of empty lines
Now it's two separate files.
* Don't preserve blank lines anymore
Makes it a little easier to copy n' paste into the Rust file.
* Start adding back cls.rs
* Add a `cls1` file for uniformity
I know this is not a very complicated blob of SMT, but now we have a 1:1
ratio of Rust functions to SMT input files.
* Convert cls1
* Generate cls8
* Generate cls8 code
* Finish generating cls16
* Generate cls32
* Finish generating cls
* Fix a bug in cls1
* Remove needless import
* Fix a variable reference in a64cs32
* Enable calls to cls* functions
* Fix return variable names for clz*
* Add clz1 SMT input
* Restore clz.rs and generate clz1
* Generate clz.rs, hopefully
But not a64clz32... don't have a file for this one?
* Rearrange some `todo!()` calls
* Remove an unnecessary `mut`
* Add extensions to the SMT files
* Add a README for encodings generation
* Remove dangling references to rsmt2
* Add a64clz32
* Generate a64clz32
* better printing/logging
* fix boolean literals, better model display
* Add rev tests that hit the solver
* Try writing rbit32 in raw SMT
Based on @mpardeshi's advice, this is just a copy of `rev32` but with an
extra extract at the top. I hope this works.
* Generate rbit32
That's the last one, hopefully!
* Run tests with timeout and retry attempt
* actually retry, longer timeout
* fix test output on failed test
* Add more encoding tests (some failing)
* more
* final rounds for clzs
* fix 64 case, add tests for cls encoding

---------

Co-authored-by: Alexa VanHattum <alexavanhattum@gmail.com>